requirements.txt
flask
adopy
pandas
numpy
pyyaml
redis
